// snaps.wms.web/src/app/externalsource/models/snaps.wms.externalsource.thparty.ts
export interface exsThirdparty { 
    orgcode: string;
    site: string;
    depot: string;
    spcarea: string;
    thtype: string;
    thbutype: string;
    thcode: string;
    thcodealt: string;
    vatcode: string;
    thname: string;
    thnamealt: string;
    addressln1: string;
    addressln2: string;
    addressln3: string;
    subdistrict: string;
    district: string;
    city: string;
    country: string;
    postcode: string;
    region: string;
    telephone: string;
    email: string;
    thgroup: string;
    thcomment: string;
    throuteformat: string;
    plandelivery: number;
    naturalloss: number;
    mapaddress: string;
    orbitsource: string;
    tflow: string;
    fileid: string;
    rowops: string;
    ermsg: string;
    dateops: Date | string | null;
    rowid: number;
}