// snaps.wms.web/src/app/externalsource/models/snaps.wms.externalsource.product.ts
export interface exsProduct { 
    orgcode: string;
    site: string;
    depot: string;
    spcarea: string;
    article: string;
    articletype: string;
    pv: number;
    lv: number;
    description: string;
    descalt: string;
    thcode: string;
    dlcall: number;
    dlcfactory: number;
    dlcwarehouse: number;
    dlcshop: number;
    dlconsumer: number;
    hdivison: string;
    hdepartment: string;
    hsubdepart: string;
    hclass: string;
    hsubclass: string;
    typemanage: string;
    unitmanage: string;
    unitdesc: string;
    unitreceipt: string;
    unitprep: string;
    unitsale: string;
    unitstock: string;
    unitweight: string;
    unitdimension: string;
    unitvolume: string;
    hucode: string;
    rtoskuofpu: number;
    rtoskuofipck: number;
    rtoskuofpck: number;
    rtoskuoflayer: number;
    rtoskuofhu: number;
    rtopckoflayer: number;
    rtolayerofhu: number;
    innaturalloss: number;
    ounaturalloss: number;
    costinbound: number;
    costoutbound: number;
    costavg: number;
    skulength: number;
    skuwidth: number;
    skuheight: number;
    skugrossweight: number;
    skuweight: number;
    skuvolume: number;
    pulength: number;
    puwidth: number;
    puheight: number;
    pugrossweight: number;
    puweight: number;
    puvolume: number;
    ipcklength: number;
    ipckwidth: number;
    ipckheight: number;
    ipckgrossweight: number;
    ipckweight: number;
    ipckvolume: number;
    pcklength: number;
    pckwidth: number;
    pckheight: number;
    pckgrossweight: number;
    pckweight: number;
    pckvolume: number;
    layerlength: number;
    layerwidth: number;
    layerheight: number;
    layergrossweight: number;
    layerweight: number;
    layervolume: number;
    hulength: number;
    huwidth: number;
    huheight: number;
    hugrossweight: number;
    huweight: number;
    huvolume: number;
    isdangerous: number;
    ishighvalue: number;
    isfastmove: number;
    isslowmove: number;
    isprescription: number;
    isdlc: number;
    ismaterial: number;
    isunique: number;
    isalcohol: number;
    istemperature: number;
    isdynamicpick: number;
    ismixingprep: number;
    isfinishgoods: number;
    isnaturalloss: number;
    isbatchno: number;
    ismeasurement: number;
    roomtype: string;
    tempmin: number;
    tempmax: number;
    alcmanage: string;
    alccategory: string;
    alccontent: string;
    alccolor: string;
    dangercategory: string;
    dangerlevel: string;
    stockthresholdmin: number;
    stockthresholdmax: number;
    spcrecvzone: string;
    spcrecvaisle: string;
    spcrecvbay: string;
    spcrecvlevel: string;
    spcrecvlocation: string;
    spcprepzone: string;
    spcdistzone: string;
    spcdistshare: string;
    spczonedelv: string;
    orbitsource: string;
    tflow: string;
    fileid: string;
    rowops: string;
    datecreate: Date | string | null;
    dateops: Date | string | null;
    ermsg: string;
    rowid: number;
    rtoipckofpck: number;
}